requirements.txt
openai>=1.40.0
boto3>=1.34.0
psycopg2-binary>=2.9.9
pgvector>=0.3.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pyyaml>=6.0.1
tiktoken>=0.7.0
click>=8.1.7
rich>=13.7.0
tenacity>=8.2.3
pytest>=8.0.0
